Fix: confirm追加

+++ phonebook/src/Person.jsx
@@ -3,8 +3,10 @@ import axios from 'axios'
 
 const Person = ({person, persons, setPersons}) => {
   const handleClick = async () => {
-    await axios.delete(`http://localhost:3001/persons/${person.id}`)
-    setPersons(persons.filter(deleted => deleted.id !== person.id))
+    if (window.confirm(`Delete ${person.name}?`)) {
+      await axios.delete(`http://localhost:3001/persons/${person.id}`)
+      setPersons(persons.filter(deleted => deleted.id !== person.id))
+    }
   }
 
   return (
